References

DataFlowTasks.DataFlowTasks — Module

moduel DataFlowTask

Create Tasks wich keep track of how data flows through it.

DataFlowTasks.LOGGER — Constant

const LOGGER::Ref{Logger}

Global Logger being used to record the events. Can be changed using setlogger!.

DataFlowTasks.SCHEDULER — Constant

const SCHEDULER::Ref{TaskGraphScheduler}

The active scheduler being used.

DataFlowTasks.TASKCOUNTER — Constant

const TASKCOUNTER::Ref{Int}

Global counter of created DataFlowTasks.

DataFlowTasks.AccessMode — Type

@enum AccessMode READ WRITE READWRITE

Describe how a DataFlowTask access its data.

DataFlowTasks.DAG — Type

struct DAG{T}

Representation of a directed acyclic graph containing nodes of type T. The list of nodes with edges coming into a node i can be retrieved using inneighbors(dag,i); similarly, the list of nodes with edges leaving from i can be retrieved using outneighbors(dag,i).

DAG is a buffered structure with a buffer of size sz_max: calling addnode! on it will block if the DAG has more than sz_max elements.

DataFlowTasks.DAG — Method

DAG{T}(sz)

Create a buffered DAG holding a maximum of s nodes of type T.

DataFlowTasks.DataFlowTask — Type

DataFlowTask(func,data,mode)

Create a task-like object similar to Task(func) which accesses data with AccessMode mode.

When a DataFlowTask is created, the elements in its data field will be checked against all other active DataFlowTask to determined if a dependency is present based on a data-flow analysis. The resulting Task will then wait on those dependencies.

A DataFlowTask behaves much like a Julia Task: you can call wait(t), schedule(t) and fetch(t) on it.

See also: @dtask, @dspawn, @dasync.

DataFlowTasks.FinishedChannel — Type

struct FinishedChannel{T} <: AbstractChannel{T}

Used to store tasks which have been completed, but not yet removed from the underlying DAG. Taking from an empty FinishedChannel will block.

DataFlowTasks.InsertionLog — Type

struct InsertionLog

Logs the execution trace of a DataFlowTask insertion.

Fields:

time_start : time the insertion began

time_finish : time the insertion finished

taskid : the thread it is inserting

tid : the thread on wich the insertion is happening

DataFlowTasks.JuliaScheduler — Type

struct JuliaScheduler{T} <: TaskGraphScheduler{T}

Implement a simple scheduling strategy which consists of delegating the DataFlowTasks to the native Julia scheduler for execution immediately after the data dependencies have been analyzed using its dag::DAG. This is the default scheduler used by DataFlowTasks.

The main advantage of this strategy is its simplicity and composability. The main disadvantage is that there is little control over how the underlying Tasks are executed by the Julia scheduler (e.g., no priorities can be passed).

Calling JuliaScheduler(sz) creates a new scheduler with an empty DAG of maximum capacity sz.

DataFlowTasks.Logger — Type

struct Logger

Contains informations on the program's progress. For thread-safety, the Logger structure uses one vector of TaskLog per thread.

DataFlowTasks.PriorityScheduler — Type

struct PriorityScheduler{T} <: TaskGraphScheduler{T}

Execute a DAG by spawning workers that take elements from the runnable channel, execute them, and put them into a finished channel to be processed by a dag_worker.

DataFlowTasks.RunnableChannel — Type

struct RunnableChannel <: AbstractChannel{DataFlowTask}

Used to store tasks which have been tagged as dependency-free, and thus can be executed. The underlying data is stored using a priority queue, with elements with a high priority being popped first.

Calling take on an empty RunnableChannel will block.

DataFlowTasks.StaticScheduler — Type

StaticScheduler{T} <: TaskGraphScheduler{T}

Like the JuliaScheduler, but requires an explicit call to execute_dag(ex) to start running the nodes in its dag (and removing them as they are completed).

Using a StaticScheduler is useful if you wish examine the underling TaskGraph before it is executed.

DataFlowTasks.Stop — Type

struct Stop

Singleton type used to safely interrupt a task reading from an `AbstractChannel.

DataFlowTasks.TaskGraph — Type

const TaskGraph = DAG{DataFlowTask}

A directed acyclic graph of DataFlowTasks.

DataFlowTasks.TaskGraphScheduler — Type

abstract type TaskGraphScheduler

Structures implementing a strategy to evaluate a DAG.

Concrete subtypes are expected to contain a dag::DAG field for storing the task graph, and a finished::AbstractChannel field to keep track of completed tasks. The interface requires the following methods:

-spawn(t,sch) -schedule(t,sch)

See also: JuliaScheduler, PriorityScheduler, StaticScheduler

DataFlowTasks.TaskLog — Type

struct TaskLog

Logs the execution trace of a DataFlowTask.

Fields:

tag : task id in DAG

time_start : time the task started running

time_finish : time the task finished running

tid : thread on wich the task ran

inneighbors : vector of incoming neighbors in DAG

label : a string used for displaying and/or postprocessing tasks

DataFlowTasks.access_mode — Method

access_mode(t::DataFlowTask[,i])

How t accesses its data.

See: AccessMode

DataFlowTasks.addedge! — Method

addedge!(dag,i,j)

Add (directed) edge connecting node i to node j in the dag.

DataFlowTasks.addedge_transitive! — Method

addedge_transitive!(dag,i,j)

Add edge connecting nodes i and j if there is no path connecting them already.

DataFlowTasks.addnode! — Method

addnode!(dag,(k,v)::Pair[, check=false])
addnode!(dag,k[, check=false])

Add a node to the dag. If passed only a key k, the value v is initialized as empty (no edges added). The check flag is used to indicate if a data flow analysis should be performed to update the dependencies of the newly inserted node.

DataFlowTasks.adjacency_matrix — Method

adjacency_matrix(dag)

Construct the adjacency matrix of dag.

DataFlowTasks.consume_runnable — Function

consume_runnable(runnable,nt,background=false)

Spawn nt = Threads.nthreads() workers that will consume tasks from runnable and execute them. If background=true the main thread (Threads.threadid()==1) is not used, and only nt-1 tasks are spawned.

DataFlowTasks.criticalpath — Function

criticalpath(logger) --> path

Finds the critical path of the logger's DAG

DataFlowTasks.dagplot — Function

dagplot(logger=getlogger())

Plot the dag in DOT format

DataFlowTasks.data — Method

data(t::DataFlowTask[,i])

Data accessed by t.

DataFlowTasks.data_dependency — Method

data_dependency(t1::DataFlowTask,t1::DataFlowTask)

Determines if there is a data dependency between t1 and t2 based on the data they read from and write to.

DataFlowTasks.enable_debug — Function

enable_debug(mode = true)

If mode is true (the default), enable debug mode: errors inside tasks will be shown.

DataFlowTasks.enable_log — Function

enable_log(mode = true)

If mode=true, information regarding the DataFlowTasks will be logged in the current logger.

See also: getlogger, setlogger!, TaskLog.

DataFlowTasks.execute_dag — Method

execute_dag(sch::StaticScheduler)

Execute all the nodes in the task graph, removing them from the dag as they are completed. This function waits for the dag to be emptied before returning.

DataFlowTasks.finished_to_runnable — Method

finished_to_runnable(dag,runnable,finished)

Worker which takes nodes from finished, remove them from the dag, and put! new nodes in runnable if they become dependency-free.

DataFlowTasks.force_sequential — Function

force_sequential(mode = true)

If mode is true, enable sequential mode: no tasks are created and scheduled, code is simply run as it appears in the sources. In effect, this makes @dspawn a no-op.

By default, sequential mode is disabled when the program starts.

DataFlowTasks.getlogger — Method

getlogger()

Return the global logger.

DataFlowTasks.getscheduler — Method

getscheduler(sch)

Return the active scheduler.

DataFlowTasks.has_edge — Method

has_edge(dag,i,j)

Check if there is an edge connecting i to j.

DataFlowTasks.inneighbors — Method

inneighbors(dag,i)

List of predecessors of i in dag.

DataFlowTasks.isconnected — Method

isconnected(dag,i,j)

Check if there is path in dag connecting i to j.

DataFlowTasks.loggertodot — Function

loggertodot(logger)  --> dagstring

Return a string in the DOT format representing the underlying graph in logger and to be plotted by GraphViz with Graph(loggertodot())

DataFlowTasks.memory_overlap — Method

memory_overlap(di,dj)

Determine if data di and dj have overlapping memory in the sense that mutating di can change dj (or vice versa). This function is used to build the dependency graph between DataFlowTasks.

A generic version is implemented returning true (but printing a warning). Users should overload this function for the specific data types used in the arguments to allow for appropriate inference of data dependencies.

DataFlowTasks.memory_overlap — Method

memory_overlap(di::Array,dj::Array)
memory_overlap(di::SubArray,dj::Array)
memory_overlap(di::Array,dj::SubArray)

When both di and dj are Arrays of bitstype, compare their addresses. If one is of type SubArray, compare the parent.

DataFlowTasks.memory_overlap — Method

memory_overlap(di::SubArray,dj::SubArray)

First compare their parents. If they are the same, compare the indices in the case where the SubArrays have the same dimension.

DataFlowTasks.num_edges — Method

num_edges(dag::DAG)

Number of edges in the DAG.

DataFlowTasks.num_nodes — Method

num_nodes(dag::DAG)

Number of nodes in the DAG.

DataFlowTasks.outneighbors — Method

outneighbors(dag,i)

List of successors of j in dag.

DataFlowTasks.plot — Function

plot(logger; categories)

Plot DataFlowTasks logger labeled informations with categories.

Entries in categories define how to group tasks in categories for plotting. Each entry can be:

a String: in this case, all tasks having labels in which the string occurs are grouped together. The string is also used as a label for the category itself.

a String => Regex pair: in this case, all tasks having labels matching the regex are grouped together. The string is used as a label for the category itself.

Example

using CairoMakie
using DataFlowTasks
using DataFlowTasks: plot, resetlogger!, sync

init!(A) = (A .= rand())                # Write
mutate!(A) = (A .= exp.(sum(A).^2).^2)  # Read/Write
get(A,B) = A+B                          # Read
function work(A, B)
    @dspawn init!(@W(A))      label="init A"
    @dspawn init!(@W(B))      label="init B"
    @dspawn mutate!(@RW(A))   label="mutate A"
    @dspawn mutate!(@RW(B))   label="mutate B"
    @dspawn get(@R(A), @R(B)) label="read A,B"
    sync()
end

# Context
A = ones(2000, 2000)
B = ones(2000, 2000)

# Compilation
# run your code once to avoid seeing artifacts related to compilation in your logged data
work(copy(A), copy(B))

# Start "real" profiling work in a clean environment
# - reset the internal logger state to discard data collected during previous runs
# - start from a clean memory state. If garbage collection happens during the
#   run, we'll know it's triggered by the real workload and the visualization will
#   highlight its impact.
resetlogger!()
GC.gc()

# Real Work
work(A, B)

# Logger Visualization
plot(categories=["init", "read", "work on B" => r"B$"])

DataFlowTasks.priority — Method

priority(t::DataFlowTask)

Function called to determine the scheduled priority of t. The default imlementation simply retuns t.priority.

DataFlowTasks.remove_node! — Method

remove_node!(dag::DAG,i)

Remove node i and all of its edges from dag.

DataFlowTasks.resetcounter! — Method

resetcounter!()

Reset the TASKCOUNTER to 0.

DataFlowTasks.resetlogger! — Function

resetlogger(logger=getlogger())

Clear the logger's memory, logging states, and reset environnement for new measures.

DataFlowTasks.savedag — Method

savedag(filepath, graph)

Save svg dag image in filepath

DataFlowTasks.setlogger! — Method

setlogger!(l::Logger)

Set the global (default) logger to l.

DataFlowTasks.setscheduler! — Method

setscheduler!(r)

Set the active scheduler to r.

DataFlowTasks.start_dag_worker — Function

start_dag_worker(sch)

Start a forever-running task associated with sch which takes nodes from finished and removes them from the dag. The task blocks if finished is empty.

DataFlowTasks.sync — Function

sync([sch::TaskGraphScheduler])

Wait for all nodes in sch to be finished before continuining. If called with no arguments, use the current scheduler.

DataFlowTasks.update_edges! — Method

update_edges!(dag::DAG,i)

Perform the data-flow analysis to update the edges of node i. Both incoming and outgoing edges are updated.

DataFlowTasks.with_scheduler — Method

with_scheduler(f,sch)

Run f, but push DataFlowTasks to the scheduler dag in sch instead of the default dag.

DataFlowTasks.@dasync — Macro

@dasync expr [kwargs...]

Like @dspawn, but schedules the task to run on the current thread.

See also:

@dspawn, @dtask

DataFlowTasks.@dspawn — Macro

@dspawn expr [kwargs...]

Create a DataFlowTask to execute the code given by expr, and schedule it to run on any available thread. The code in expr should be annotated with @R, @W and/or @RW tags in order to indicate how it accesses data (see examples below). This information is is then used to automatically infer task dependencies.

Supported keyword arguments:

label: provide a label to identify the task. This is useful when logging scheduling information;

priority: inform the scheduler about the relative priority of the task. This information is not (yet) leveraged by the default scheduler.

See also:

@dtask, @dasync

Examples:

Below are 3 equivalent ways to create the same DataFlowTask, which expresses a Read-Write dependency on C and Read dependencies on A and B

using LinearAlgebra
A = rand(10, 10)
B = rand(10, 10)
C = rand(10, 10)
α, β = (100.0, 10.0)

# Option 1: annotate arguments in a function call
@dspawn mul!(@RW(C), @R(A), @R(B), α, β)

# Option 2: specify data access modes in the code block
@dspawn begin
   @RW C
   @R  A B
   mul!(C, A, B, α, β)
end

# Option 3: specify data access modes after the code block
# (i.e. alongside keyword arguments)
@dspawn mul!(C, A, B, α, β) @RW(C) @R(A,B)

Here is a more complete example, demonstrating a full computation involving 2 different tasks.

using DataFlowTasks

A = rand(5)

# create a task with WRITE access mode to A
# and label "writer"
t1 = @dspawn begin
    @W A
    sleep(1)
    fill!(A,0)
    println("finished writing")
end  label="writer"

# create a task with READ access mode to A
t2 = @dspawn begin
    @R A
    println("I automatically wait for `t1` to finish")
    sum(A)
end  priority=1

fetch(t2) # 0

# output

finished writing
I automatically wait for `t1` to finish
0.0

Note that in the example above t2 waited for t1 because it read a data field that t1 accessed in a writable manner.

DataFlowTasks.@dtask — Macro

@dtask expr [kwargs...]

Create a DataFlowTask to execute expr, where data have been tagged to specify how they are accessed. Note that the task is not automatically scheduled for execution.

See @dspawn for information on how to annotate expr to specify data dependencies, and a list of supported keyword arguments.

See also:

@dspawn, @dasync