References

DataFlowTasks.DataFlowTasks — Module

module DataFlowTask

Create Tasks which can keep track of how data flows through it.

DataFlowTasks.LOGINFO — Constant

const LOGINFO::Ref{LogInfo}

Global LogInfo being used to record the events. Can be changed using _setloginfo!.

DataFlowTasks.SCHEDULER — Constant

const SCHEDULER::Ref{TaskGraphScheduler}

The active scheduler being used.

DataFlowTasks.TASKCOUNTER — Constant

const TASKCOUNTER::Ref{Int}

Global counter of created DataFlowTasks.

DataFlowTasks.AccessMode — Type

@enum AccessMode READ WRITE READWRITE

Describe how a DataFlowTask access its data.

DataFlowTasks.DAG — Type

struct DAG{T}

Representation of a directed acyclic graph containing nodes of type T. The list of nodes with edges coming into a node i can be retrieved using inneighbors(dag,i); similarly, the list of nodes with edges leaving from i can be retrieved using outneighbors(dag,i).

DAG is a buffered structure with a buffer of size sz_max: calling addnode! on it will block if the DAG has more than sz_max elements.

DataFlowTasks.DAG — Method

DAG{T}(sz)

Create a buffered DAG holding a maximum of sz nodes of type T.

DataFlowTasks.DataFlowTask — Type

DataFlowTask(func,data,mode)

Create a task-like object similar to Task(func) which accesses data with AccessMode mode.

When a DataFlowTask is created, the elements in its data field will be checked against all other active DataFlowTask to determined if a dependency is present based on a data-flow analysis. The resulting Task will then wait on those dependencies.

A DataFlowTask behaves much like a Julia Task: you can call wait(t), schedule(t) and fetch(t) on it.

See also: @dtask, @dspawn, @dasync.

DataFlowTasks.FinishedChannel — Type

struct FinishedChannel{T} <: AbstractChannel{T}

Used to store tasks which have been completed, but not yet removed from the underlying DAG. Taking from an empty FinishedChannel will block.

DataFlowTasks.InsertionLog — Type

struct InsertionLog

Logs the execution trace of a DataFlowTask insertion.

Fields:

time_start : time the insertion began

time_finish : time the insertion finished

taskid : the task it is inserting

tid : the thread on which the insertion is happening

DataFlowTasks.JuliaScheduler — Type

struct JuliaScheduler{T} <: TaskGraphScheduler{T}

Implements a simple scheduling strategy which consists of delegating the DataFlowTasks to the native Julia scheduler for execution immediately after the data dependencies have been analyzed using its dag::DAG. This is the default scheduler used by DataFlowTasks.

The main advantage of this strategy is its simplicity and composability. The main disadvantage is that there is little control over how the underlying Tasks are executed by the Julia scheduler (e.g., no priorities can be passed).

Calling JuliaScheduler(sz) creates a new scheduler with an empty DAG of maximum capacity sz.

DataFlowTasks.LogInfo — Type

struct LogInfo

Contains informations on the program's progress. For thread-safety, the LogInfo structure uses one vector of TaskLog per thread.

You can visualize and postprocess a LogInfo using GraphViz.Graph and Makie.plot.

DataFlowTasks.Stop — Type

struct Stop

Singleton type used to safely interrupt a task reading from an AbstractChannel.

DataFlowTasks.TaskGraph — Type

const TaskGraph = DAG{DataFlowTask}

A directed acyclic graph of DataFlowTasks.

DataFlowTasks.TaskGraphScheduler — Type

abstract type TaskGraphScheduler

Structures implementing a strategy to evaluate a DAG.

Concrete subtypes are expected to contain a dag::DAG field for storing the task graph, and a finished::AbstractChannel field to keep track of completed tasks. The interface requires the following methods:

-spawn(t,sch) -schedule(t,sch)

See also: JuliaScheduler

DataFlowTasks.TaskLog — Type

struct TaskLog

Logs the execution trace of a DataFlowTask.

Fields:

tag : task id in DAG

time_start : time the task started running

time_finish : time the task finished running

tid : thread on which the task ran

inneighbors : vector of incoming neighbors in DAG

label : a string used for displaying and/or postprocessing tasks

DataFlowTasks._getloginfo — Method

_getloginfo()

Return the active logger.

DataFlowTasks._setloginfo! — Method

_setloginfo!(l::LogInfo)

Set the active logger to l.

DataFlowTasks.access_mode — Method

access_mode(t::DataFlowTask[,i])

How t accesses its data.

See: AccessMode

DataFlowTasks.addedge! — Method

addedge!(dag,i,j)

Add (directed) edge connecting node i to node j in the dag.

DataFlowTasks.addedge_transitive! — Method

addedge_transitive!(dag,i,j)

Add edge connecting nodes i and j if there is no path connecting them already.

DataFlowTasks.addnode! — Method

addnode!(dag,(k,v)::Pair[, check=false])
addnode!(dag,k[, check=false])

Add a node to the dag. If passed only a key k, the value v is initialized as empty (no edges added). The check flag is used to indicate if a data flow analysis should be performed to update the dependencies of the newly inserted node.

DataFlowTasks.capacity — Method

capacity(dag)

The maximum number of nodes that dag can contain.

DataFlowTasks.data — Method

data(t::DataFlowTask[,i])

Data accessed by t.

DataFlowTasks.data_dependency — Method

data_dependency(t1::DataFlowTask,t1::DataFlowTask)

Determines if there is a data dependency between t1 and t2 based on the data they read from and write to.

DataFlowTasks.enable_debug — Function

enable_debug(mode = true)

If mode is true (the default), enable debug mode: errors inside tasks will be shown.

DataFlowTasks.enable_log — Function

enable_log(mode = true)

If mode is true (the default), logging is enabled throug the @log macro. Calling enable_log(false) will de-activate logging at compile time to avoid any possible overhead.

Note that changing the log mode at runtime will may invalidate code, possibly triggering recompilation.

See also: @log, with_logging

DataFlowTasks.force_linear_dag — Function

force_linear_dag(mode=false)

If mode is true, nodes are added to the DAG in a linear fashion, i.e. the DAG connects node i to node i+1. This is useful for debugging purposes.

DataFlowTasks.force_sequential — Function

force_sequential(mode = true)

If mode is true, enable sequential mode: no tasks are created and scheduled, code is simply run as it appears in the sources. In effect, this makes @dspawn a no-op.

By default, sequential mode is disabled when the program starts.

See also: force_linear_dag.

DataFlowTasks.getscheduler — Method

getscheduler()

Return the active scheduler.

DataFlowTasks.has_edge — Method

has_edge(dag,i,j)

Check if there is an edge connecting i to j.

DataFlowTasks.inneighbors — Method

inneighbors(dag,i)

List of predecessors of i in dag.

DataFlowTasks.isconnected — Method

isconnected(dag,i,j)

Check if there is a path in dag connecting i to j.

DataFlowTasks.memory_overlap — Method

memory_overlap(di,dj)

Determine if data di and dj have overlapping memory in the sense that mutating di can change dj (or vice versa). This function is used to build the dependency graph between DataFlowTasks.

A generic version is implemented returning true (but printing a warning). Users should overload this function for the specific data types used in the arguments to allow for appropriate inference of data dependencies.

DataFlowTasks.memory_overlap — Method

memory_overlap(di::AbstractArray,dj::AbstractArray)

Try to determine if the arrays di and dj have overlapping memory.

When both di and dj are Arrays of bitstype, simply compare their addresses. Otherwise, compare their parents by default.

When both di and dj are SubArrays we compare the actual indices of the SubArrays when their parents are the same (to avoid too many false positives).

DataFlowTasks.nodes — Method

nodes(dag::DAG)

Return an iterator over the nodes of dag.

DataFlowTasks.num_edges — Method

num_edges(dag::DAG)

Number of edges in the DAG.

DataFlowTasks.num_nodes — Method

num_nodes(dag::DAG)

Number of nodes in the DAG.

DataFlowTasks.outneighbors — Method

outneighbors(dag,i)

List of successors of j in dag.

DataFlowTasks.remove_node! — Method

remove_node!(dag::DAG,i)

Remove node i and all of its edges from dag.

DataFlowTasks.restart_scheduler! — Method

restart_scheduler!([sch])

Interrupt all tasks in sch and remove them from the underlying DAG.

This function is useful to avoid having to restart the REPL when a task in sch errors.

DataFlowTasks.savedag — Function

DataFlowTasks.savedag(filepath, graph)

Save graph as an SVG image at filepath. This requires GraphViz to be available.

DataFlowTasks.setscheduler! — Method

setscheduler!(sch)

Set the active scheduler to sch.

DataFlowTasks.start_dag_worker — Function

start_dag_worker(sch)

Start a task associated with sch which takes nodes from its finished queue and removes them from the dag. The task blocks if finished is empty.

DataFlowTasks.sync — Function

sync([sch::TaskGraphScheduler])

Wait for all nodes in sch to be finished before continuining. If called with no arguments, use the current scheduler.

DataFlowTasks.update_edges! — Method

update_edges!(dag::DAG,i)

Perform the data-flow analysis to update the edges of node i. Both incoming and outgoing edges are updated.

DataFlowTasks.with_logging! — Method

with_logging!(f,l::LogInfo)

Similar to with_logging, but append events to l.

DataFlowTasks.with_logging — Method

with_logging(f) --> f(),loginfo

Execute f() and log DataFlowTasks into the loginfo object.

Examples:

using DataFlowTasks

A,B = zeros(2), ones(2);

out,loginfo = DataFlowTasks.with_logging() do
    @dspawn fill!(@W(A),1)
    @dspawn fill!(@W(B),1)
    res = @dspawn sum(@R(A)) + sum(@R(B))
    fetch(res)
end

#

out

See also: LogInfo

DataFlowTasks.with_scheduler — Method

with_scheduler(f,sch)

Run f, but push DataFlowTasks to the scheduler dag in sch instead of the default dag.

DataFlowTasks.@dasync — Macro

@dasync expr [kwargs...]

Like @dspawn, but schedules the task to run on the current thread.

See also:

@dspawn, @dtask

DataFlowTasks.@dspawn — Macro

@dspawn expr [kwargs...]

Create a DataFlowTask to execute the code given by expr, and schedule it to run on any available thread. The code in expr should be annotated with @R, @W and/or @RW tags in order to indicate how it accesses data (see examples below). This information is is then used to automatically infer task dependencies.

Supported keyword arguments:

label: provide a label to identify the task. This is useful when logging scheduling information;

priority: inform the scheduler about the relative priority of the task. This information is not (yet) leveraged by the default scheduler.

See also: @dtask, @dasync

Examples:

Below are 3 equivalent ways to create the same DataFlowTask, which expresses a Read-Write dependency on C and Read dependencies on A and B

using DataFlowTasks,LinearAlgebra
A = ones(5, 5)
B = ones(5, 5)
C = zeros(5, 5)
α, β = (1, 0)

# Option 1: annotate arguments in a function call
@dspawn mul!(@RW(C), @R(A), @R(B), α, β)

# Option 2: specify data access modes in the code block
@dspawn begin
   @RW C
   @R  A B
   mul!(C, A, B, α, β)
end

# Option 3: specify data access modes after the code block
# (i.e. alongside keyword arguments)
res = @dspawn mul!(C, A, B, α, β) @RW(C) @R(A,B)

fetch(res) # a 5×5 matrix of 5.0

Here is a more complete example, demonstrating a full computation involving 2 different tasks.

using DataFlowTasks

A = rand(5)

# create a task with WRITE access mode to A
# and label "writer"
t1 = @dspawn begin
    @W A
    sleep(1)
    fill!(A,0)
    println("finished writing")
end  label="writer"

# create a task with READ access mode to A
t2 = @dspawn begin
    @R A
    println("I automatically wait for `t1` to finish")
    sum(A)
end  priority=1

fetch(t2) # 0

# output

finished writing
I automatically wait for `t1` to finish
0.0

Note that in the example above t2 waited for t1 because it read a data field that t1 accessed in a writable manner.

DataFlowTasks.@dtask — Macro

@dtask expr [kwargs...]

Create a DataFlowTask to execute expr, where data have been tagged to specify how they are accessed. Note that the task is not automatically scheduled for execution.

See @dspawn for information on how to annotate expr to specify data dependencies, and a list of supported keyword arguments.

See also: @dspawn, @dasync

DataFlowTasks.@log — Macro

DataFlowTasks.@log expr --> LogInfo

Execute expr and return a LogInfo instance with the recorded events.

Warning

The returned LogInfo instance may be incomplete if block returns before all DataFlowTasks spawened inside of it are completed. Typically expr should fetch the outcome before returning to properly benchmark the code that it runs (and not merely the tasks that it spawns).

See also: with_logging, with_logging!

DataFlowTasks.@using_opt — Macro

DataFlowTasks.@using_opt pkgnames

Load pkgnames from optional dependencies.

Warning

This feature is experimental and might break in the future. It may be useful for quick experiments, but adding GraphViz or Makie to a full-fledged environment should be preferred when possible.

Examples:

using DataFlowTasks
DataFlowTasks.@using_opt GraphViz

GraphViz.Graph — Method

GraphViz.Graph(log_info::LogInfo)

Produce a GraphViz.Graph representing the DAG of tasks collected in log_info.

See also: DataFlowTasks.@log

DataFlowTasks_GraphViz_Ext.loggertodot — Method

loggertodot(logger)  --> dagstring

Return a string in the DOT format representing the underlying graph in logger and to be plotted by GraphViz with Graph(loggertodot())

MakieCore.plot — Method

plot(log_info; categories)

Plot DataFlowTasks log_info labeled informations with categories.

Entries in categories define how to group tasks in categories for plotting. Each entry can be:

a String: in this case, all tasks having labels in which the string occurs are grouped together. The string is also used as a label for the category itself.

a String => Regex pair: in this case, all tasks having labels matching the regex are grouped together. The string is used as a label for the category itself.

See the documentation for more information on how to profile and visualize DataFlowTasks.